rename library name
