--- KEXtentions/src/main/java/ru/nightgoat/kextentions/ThrowableExt.kt
@@ -1,3 +1,0 @@
-package ru.nightgoat.kextentions
-
-fun Throwable.getNameAndMessage() = "KEXception: $this, message: $message"
