--- KEXtentions/src/main/java/ru/nightgoat/kextentions/utils/logging/AndroidLogger.kt
@@ -1,11 +1,0 @@
-package ru.nightgoat.kextentions.utils.logging
-
-import android.annotation.SuppressLint
-import android.util.Log
-
-object AndroidLogger : ILogger {
-    @SuppressLint("LogNotTimber")
-    override fun loggE(message: String, tag: String?, e: Throwable?) {
-        Log.e(tag, message, e)
-    }
-}
