--- KEXtentions/src/main/java/ru/nightgoat/kextentions/utils/logging/ILogger.kt
@@ -1,5 +1,0 @@
-package ru.nightgoat.kextentions.utils.logging
-
-interface ILogger {
-    fun loggE(message: String, tag: String?, e: Throwable? = null)
-}
