+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "ru.nightgoat.kextentions"
+        applicationId "ru.nightgoat.kextensions"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
@@ -43,7 +43,7 @@ android {
 
 dependencies {
 
-    api project(':KEXtentions')
+    api project(':KEXtensions')
     testImplementation "org.mockito:mockito-core:3.7.7"
     testImplementation 'junit:junit:4.13.2'
 }
