+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = "KEXTentions"
+rootProject.name = "KEXTensions"
 include ':app'
-include ':KEXtentions'
+include ':KEXtensions'
